chatbot/requirements.txt
openai==1.3.0
langchain==0.0.300
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv
google-generativeai
langchain
python-multipart==0.0.6
